modify branches struct
